+++ webapp/src/main/java/rocks/metaldetector/service/user/events/OnRegistrationCompleteEvent.java
@@ -1,8 +1,9 @@
-package rocks.metaldetector.service.user;
+package rocks.metaldetector.service.user.events;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.context.ApplicationEvent;
+import rocks.metaldetector.service.user.UserDto;
 
 @Getter
 @Setter
